:test_tube: Add multiple failing test
Adding multiple failing test to test the functionality of the command parser. Sadly all of them fail. It is to be determined if they fail because they are written wrongly or because the parser fails to parse the respective commands. Tests need also be reviewed if they are useful and correct.

Signed-off-by: Sören W <galaxymc.de@gmail.com>