:sparkles: :white_check_mark: Adding support and tests for bool arguments
Adding support for bool arguments.
Adding tests to test (duh) the functionality of bools isolated as well as with sounding arguments.

The tests are all passing.

I'm not sure about my code style here, so I didn't want to commit it to master directly (you shouldn't do this anyway, but it has been the current workflow).

Signed-off-by: Sören Wicke Iurgel <soeren.wicke.iurgel@gmail.com>